disable increment, decrement of ID Number input
Change type of ID Number input to 'text' and to disable increment,
decrement buttons; then use the onInput event to prevent non-numeric
characters from being entered.
This is to disallow inputs like '-123', remove the up and down arrow
as well as the scroll up-down to increase, decrease
the value of the input.