add the hover effect to the social icons at footer
Make the social icons at the footer have a hover effect that changes
the background color to orange (be the same as the color of text when
hover) in the shape of a circle (surrounding the icon).
The reason for the circle is because the page also has a circle
at the center under the title of every pages (under "HOME", "ROOM LIST",
"BOOKING LIST", etc).