allow running BE-FE, add .env
Add package 'concurrently' to run both BE-FE in different ports;
update .gitignore to ignore .env file and add .env.example file;
Now if you want to run the project, you need to create a .env file
in the same folder as .env.example. If want to run both BE and FE,
run 'npm run dev' in the 'server' folder.