feat: FE get data from BE and display
Modify the path and the MockData in the BE side to satisfy the FE
side request, then successfully display the data in the FE side.
Add the .env.example file to the BE side to show the environment, also
add .env to the .gitignore.
Install 'CORS' and use it in the BE side to allow fetching data from FE.